+++ holoviews/plotting/bokeh/plot.py
@@ -32,9 +32,8 @@ from ..plot import (
 from ..util import attach_streams, displayable, collate
 from .callbacks import LinkCallback
 from .util import (
-    TOOL_TYPES, filter_toolboxes, make_axis, update_shared_sources,
-    empty_plot, decode_bytes, theme_attr_json, cds_column_replace,
-    get_default
+    filter_toolboxes, make_axis, update_shared_sources, empty_plot,
+    decode_bytes, theme_attr_json, cds_column_replace, get_default
 )
